LayoutTests/fast/canvas/canvas-arc-360-winding-expected.txt

This tests canvas full arc fill with nonzero winding rule. Eight green concentric thick circumferences should be displayed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS data[0] is 0
PASS data[1] is 0
PASS data[2] is 0
PASS data[0] is 0
PASS data[1] is 255
PASS data[2] is 0
PASS data[0] is 0
PASS data[1] is 0
PASS data[2] is 0
PASS successfullyParsed is true

TEST COMPLETE
